chore: deduplicate `is-land` script tags in prod
Instead of relying on `svelte:head` for production, this patch makes the
plugin directly inject the `is-land` script tag into the rendered HTML.
Since we already do replacements on the HTML, this is not a problem.

For dev mode, `svelte:head` is still used due to the lack of a way to
directly inject a script tag. SvelteKit doesn't use Vite to transform
the `index.html`, so using `transformIndexHtml` isn't an option.